pyproject.toml
[project]
name = "payment-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "faststream[rabbit]>=0.6.7",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
